requirements.txt
fastapi
uvicorn
jinja2
tortoise-orm
pydantic
